Externalizing LogicTypeSelector from ExpressionContainer
