+++ kie-wb-common-react/boxed-expression-editor/src/components/LogicTypeSelector/LogicTypeSelector.css
@@ -14,19 +14,23 @@
  * limitations under the License.
  */
 
-.expression-container .expression-container-box {
-  border-style: ridge;
-  border-color: #39A5DC;
+.expression-container .logic-type-selector {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  display: -webkit-box;
+  display: -moz-box;
+  display: -ms-flexbox;
+  display: -webkit-flex;
+  min-width: 15em;
+  min-height: 4em;
 }
 
-.expression-container .context-menu-container {
-  position: absolute;
-  z-index: 10000;
+.expression-container .logic-type-selector.logic-type-not-present {
+  cursor: pointer;
 }
 
-.expression-container > .expression-container-box,
-.expression-container .expression-type {
-  font-size: smaller;
-  font-style: italic;
-  color: gray;
+.expression-container .logic-type-selector.logic-type-selected {
+  cursor: default;
+  padding: 1em;
 }
